sftp_test/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.2
Django==4.2.3
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.14.0
paramiko==3.2.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
